finish archived posts list before cleaning up

+++ src/app/api/admin/deletePost/route.ts
@@ -4,13 +4,14 @@ import { NextRequest } from "next/server";
 
 export async function GET(request: NextRequest) {
     const param = request.nextUrl.searchParams.get("postId")
-    if (!param) {
-        return Response.json({ message: "Please provide a post id", ok: false }, {
+    const route = request.nextUrl.searchParams.get("route")
+    if (!param || !route) {
+        return Response.json({ message: "Please provide a post id and a route", ok: false }, {
             status: 400
         })
     }
     try {
-        const docRef = doc(collection(db, "posts"), param)
+        const docRef = doc(collection(db, route), param)
         await deleteDoc(docRef)
         return Response.json({ "message": `post with Id:${param} was deleted successfully`, ok: true}, {
             status: 200
